feat(tracking): add updating tracking info api

+++ business-logic/src/main/java/ru/itmo/cloudtechonlogies/dto/PageTimerDTO.java
@@ -7,5 +7,5 @@ import lombok.Data;
 @Builder
 public class PageTimerDTO {
     private Integer page;
-    private Long timer;
+    private Integer timer;
 }
